+++ packages/cli/src/commands/__test__/exportPreviews.test.ts
@@ -35,4 +35,11 @@ describe("exportPreviews command", () => {
       expect(out).toMatchSnapshot();
     });
   });
+
+  describe("cli --minify", () => {
+    it("runs on templates", async () => {
+      const out = await execCli("export-previews --minify");
+      expect(out).toMatchSnapshot();
+    });
+  });
 });